Tidy up booking flows, add availability flows
