+++ packages/openactive-integration-tests/test/flows/confirm-availability/no-spaces-available/book-shared.js
@@ -1,0 +1,94 @@
+const chakram = require("chakram");
+const expect = chakram.expect;
+const pMemoize = require("p-memoize");
+
+const RequestHelper = require("../../../helpers/request-helper");
+const Logger = require("../../../helpers/logger");
+const sharedValidationTests = require("../../../shared-behaviours/validation");
+
+function performTests(dataItem) {
+  const { event: testEvent, price, name: eventName } = dataItem;
+
+  var opportunityId;
+  var offerId;
+  var sellerId;
+  var uuid;
+  var totalPaymentDue;
+
+  var c1Response;
+
+  const logger = new Logger(dataItem.title);
+
+  const testHelper = new RequestHelper(logger);
+
+  beforeAll(async function() {
+    logger.log(
+      "\n\n** Test Event **: \n\n" + JSON.stringify(testEvent, null, 2)
+    );
+
+    uuid = testHelper.uuid();
+
+    // get the getMatch in before we create the session (helps with race conditions)
+    performGetMatch();
+
+    let session = await testHelper.createScheduledSession(testEvent, {
+      sellerId
+    });
+
+    return chakram.wait();
+  });
+
+  afterAll(async function() {
+    // by the end, it should have done this already, but let's force it through if it hasn't
+
+    await testHelper.deleteScheduledSession(eventName, {
+      sellerId
+    });
+    return chakram.wait();
+  });
+
+  const performGetMatch = pMemoize(async function performGetMatch() {
+    ({ opportunityId, offerId, sellerId } = await testHelper.getMatch(
+      eventName
+    ));
+  });
+
+  const performC1 = pMemoize(async function performC1() {
+    await performGetMatch();
+
+    ({ c1Response, totalPaymentDue } = await testHelper.putOrderQuoteTemplate(
+      uuid,
+      {
+        opportunityId,
+        offerId,
+        sellerId,
+        uuid
+      }
+    ));
+  });
+
+  describe("C1", function() {
+    beforeAll(async function() {
+      await performC1();
+    });
+
+    it("should return 409 - Conflict", async function() {
+      expect(c1Response).to.have.status(409);
+    });
+
+    it("should return a OpportunityIsFullError error", async function() {
+      return expect(uResponse).to.have.json(
+        "@type",
+        "CancellationNotPermittedError"
+      );
+    });
+
+    sharedValidationTests.shouldBeValidResponse(() => c1Response.body, "C1", {
+      // validationMode: "C1Response",
+    });
+  });
+}
+
+module.exports = {
+  performTests
+};
