+++ packages/openactive-integration-tests/test/setup.js
@@ -1,0 +1,3 @@
+window.ROOT_DIR = __dirname;
+
+process.env["NODE_TLS_REJECT_UNAUTHORIZED"] = 0;
